api/requirements.txt
aiohttp
aiohttp-retry
aiosignal
annotated-types
anyio
argcomplete
asgiref
attrs
Babel
beautifulsoup4
bleach
boto3
botocore
cachetools
certifi
cffi
cfn-flip
charset-normalizer
click
colorama
coreapi
coreschema
cryptography
curlify
distlib
distro
Django
django-autoslug
django-cities-light
django-cors-headers
django-environ
django-js-asset
django-restframework
django-storages
djangorestframework
drf-yasg
durationpy
filelock
frozenlist
gunicorn
h11
hjson
httpcore
httplib2
httpx
idna
inflection
itypes
Jinja2
jmespath
kappa
MarkupSafe
multidict
packaging
paypal-checkout-serversdk
paypalhttp
paypalrestsdk
pillow
placebo
platformdirs
progressbar2
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycountry
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
pyOpenSSL
pyparsing
python-amazon-paapi
python-dateutil
python-slugify
python-utils
pytz
PyYAML
requests
rsa
s3transfer
serverless-wsgi
setuptools
six
sniffio
soupsieve
sqlparse
stripe
text-unidecode
toml
tqdm
troposphere
twilio
typing_extensions
tzdata
Unidecode
uritemplate
urllib3
virtualenv
waitress
webencodings
Werkzeug
wheel
yarl
zappa
